Only delete files older than 5min

+++ util/configv3/load_config.go
@@ -7,6 +7,7 @@ import (
 	"math"
 	"os"
 	"path/filepath"
+	"time"
 
 	"code.cloudfoundry.org/cli/command/translatableerror"
 	"golang.org/x/crypto/ssh/terminal"
@@ -183,8 +184,16 @@ func removeOldTempConfigFiles() error {
 	}
 
 	for _, oldTempFileName := range oldTempFileNames {
+		fi, err := os.Lstat(oldTempFileName)
+		// ignore if file doesn't exist anymore due to race conditions if multiple cli commands are running in parallel
+		if err != nil && !errors.Is(err, os.ErrNotExist) {
+			return err
+		}
+		// only delete old orphans which are not caught by the signal handler in WriteConfig
+		if fi.ModTime().After(time.Now().Add(-5 * time.Minute)) {
+			continue
+		}
 		err = os.Remove(oldTempFileName)
-		// ignore if file doesn't exist anymore due to race conditions if multiple cli commands are running
 		if err != nil && !errors.Is(err, os.ErrNotExist) {
 			return err
 		}
